pyproject.toml
[tool.poetry]
name = "version-store-backend"
version = "0.1.0"
description = ""
authors = ["Fredrik Stockman <fredrik@versionlens.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
strawberry-graphql-django = "^0.9.2"
django-cors-headers = "^3.13.0"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
